05: Part 2
I'm so sure there was a "smart" way to do this. But I implemented a
"range-based" lookup, which returns output ranges.

So, for each input range, I look it up in the map, getting back a list
of "output ranges". Then I can look up each of those output ranges in
the next map. And so on...

Working out the overlaps or lack thereof was brain melty.

This was HARD for a week-day day 5!