10: Refactor
This is literally driving me mad. What I'm trying to do is trace the
path and keep track of the cells on the "left" and "right" of it. My
lack of understanding of the language is fighting me at every stage.

This still gets the right answer for part 1, and I think it's getting
the correct cells for "side_a" and "side_b".