18: Part 2 - polygon area with fudge?
We have to add half of the perimeter to add the extra space for each
discrete block - but this "undercounts" the external corners (?) - I
don't really get how this translates to a "+1", but it works on my
input.

Maybe it's because we're guaranteed that a closed shape has this
property?