WIP: 21: Print out some multi-grid
Still not sure how to do part 2. I assume that there are 2 "types" of
tiles - "odd" ones and "even" ones, and the whole space will be filled
with tiles matching those patterns. So it might be possible to figure
out how many "complete" tiles fit in 26501365 steps, and then figure out
the partials at the edge? Even that edge is going to be _huge_ though.

I can't figure out a "more analytical" way than that.