22: Part 2: Just brute force it.
I have a vague idea of how to do this more efficiently, deriving a tree
of the whole brick-support tree, then figure out the number of nodes
which drop from the tree if you drop a node (brick).

But hey, I'm meant to be hanging out with family, and Rust + M1 is fast.